Remove validation of select_value from the constructor.
We already did part of this validation in the SIL verifier. I've added
the remaining validation there.

In theory we should be able to do this validation in the constructor,
but the way the deserializer is implemented we run into problems in
practice because we sometimes materialize dummy placeholders for uses of
values we haven't seen the definitions for (e.g. for out-of-order blocks).

This was exposed by some pass ordering changes I expect to commit
shortly.

We should really deal with how we handle these uses differently to
enable more validation in the constructors for instructions.  I'll use
rdar://problem/24761757, which I opened for this specific issue, to
track the more general issue.