Type Re-constructor: Fix a crasher for LLDB. Patch and commit message from Greg Clayton.
Code like this:

public struct MyStruct {
   private typealias IntegerType = Int
   private func Foo(i : IntegerType)
   {
       var a = Dictionary<String, IntegerType>()
       a["hello"] = 123
       print("a = \(a)")
       print("i = \(i)")
       print("done");
   }
}
func main() {
   let s = MyStruct();
   s.Foo(123)
}

main()

Will crash if we set a breakpoint on the 'print("done")' line and expand the "a" variable due to the IntegerType not
making it into the Swift module since it is private. Then we try to create the Dictionary generic type with type
list that has only "String" in it and boom.